(#5908) - bundle all browser test files
