+++ .eslintignore
@@ -3,3 +3,4 @@ tests/integration/cordova/**/*.js
 tests/performance-bundle.js
 tests/fuzzy/seedrandom.min.js
 tests/integration/utils-bundle.js
+tests/integration/tests-bundle.js
