+++ package.json
@@ -26,7 +26,8 @@
     "publish-site": "sh bin/publish-site.sh",
     "build-site": "node ./bin/build-site.js",
     "report-coverage": "npm run test-coverage && istanbul-coveralls --no-rm",
-    "build-test": "npm run build-test-utils && npm run build-perf",
+    "build-test": "npm run build-test-utils && npm run build-test-bundle && npm run build-perf",
+    "build-test-bundle": "cat tests/integration/{aa,test,browser}.*.js > tests/integration/tests-bundle.js",
     "build-test-utils": "browserify tests/integration/utils.js > tests/integration/utils-bundle.js",
     "build-perf": "browserify tests/performance/index.js > tests/performance-bundle.js",
     "set-version": "node bin/set-version.js",
@@ -73,6 +74,7 @@
     "eslint": "3.9.0",
     "express": "4.14.0",
     "express-pouchdb": "2.1.0",
+    "fileconcat": "^2.0.1",
     "find-requires": "0.2.2",
     "glob": "7.1.1",
     "http-server": "0.9.0",
