+++ tests/integration/index.html
@@ -19,41 +19,7 @@
       var assert = chai.assert;
     </script>
     <script src='utils-bundle.js'></script>
-    <script src='test.setup_global_hooks.js'></script>
-    <script src='test.ajax.js'></script>
-    <script src='test.aa.setup.js'></script>
-    <script src='test.basics.js'></script>
-    <script src='test.constructor.js'></script>
-    <script src='test.changes.js'></script>
-    <script src='test.bulk_docs.js'></script>
-    <script src='test.bulk_get.js'></script>
-    <script src='test.all_docs.js'></script>
-    <script src='test.events.js'></script>
-    <script src='test.conflicts.js'></script>
-    <script src='test.revs_diff.js'></script>
-    <script src='test.replication.js'></script>
-    <script src='test.replication_events.js'></script>
-    <script src='test.sync.js'></script>
-    <script src='test.sync_events.js'></script>
-    <script src='test.retry.js'></script>
-    <script src='test.taskqueue.js'></script>
-    <script src='test.design_docs.js'></script>
-    <script src='browser.info.js'></script>
-    <script src='test.issue221.js'></script>
-    <script src='test.issue2674.js'></script>
-    <script src='test.issue3179.js'></script>
-    <script src='test.issue3646.js'></script>
-    <script src='test.http.js'></script>
-    <script src='test.compaction.js'></script>
-    <script src='test.get.js'></script>
-    <script src='test.local_docs.js'></script>
-    <script src='test.attachments.js'></script>
-    <script src='browser.migration.js'></script>
-    <script src='test.uuids.js'></script>
-    <script src='test.slash_id.js'></script>
-    <script src='test.reserved.js'></script>
-    <script src='test.prefix.js'></script>
-    <script src='browser.worker.js'></script>
+    <script src='tests-bundle.js'></script>
     <script type="text/javascript" src="./webrunner.js"></script>
   </body>
 </html>
