+++ tests/integration/node.node-websql.js
@@ -1,7 +1,7 @@
 'use strict';
 
 if (process.env.ADAPTER === 'websql') {
-  describe('test.node-websql.js', function () {
+  describe('node.node-websql.js', function () {
     it('should run websql when we are actually testing websql', function () {
       var db = new PouchDB('testdb');
       db.adapter.should.equal('websql');
